pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "paper-fetcher"
version = "0.1.0"
description = "Multi-source academic paper search and Sci-Hub download tool"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
    "pyyaml>=6.0",
    "rapidfuzz>=3.0",
    "beautifulsoup4>=4.11",
    "scholarly>=1.7",
    "arxiv>=2.0",
    "questionary>=2.0",
    "rich>=13.0",
]

[project.scripts]
papers = "cli.main:main"

[tool.setuptools.packages.find]
include = ["cli*", "core*"]
